zkCert: path fixes due to moved scripts
