+++ packages/zk-certificates/scripts/deploy/deployZkCertificateRegistry.ts
@@ -2,8 +2,8 @@
 import { poseidonContract } from 'circomlibjs';
 import { ethers, network } from 'hardhat';
 
-import { deploySC } from '../lib/hardhatHelpers';
-import { overwriteArtifact } from '../lib/helpers';
+import { deploySC } from '../../lib/hardhatHelpers';
+import { overwriteArtifact } from '../../lib/helpers';
 
 /**
  * Deploys a contract that everyone can use to submit encrypted Data for on-chain storage.
