+++ packages/zk-certificates/scripts/merkle/merkleTreeGenerator.ts
@@ -1,7 +1,7 @@
 /* Copyright (C) 2023 Galactica Network. This file is part of zkKYC. zkKYC is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version. zkKYC is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details. You should have received a copy of the GNU General Public License along with this program. If not, see <https://www.gnu.org/licenses/>. */
 import { buildPoseidon } from 'circomlibjs';
 
-import { MerkleTree } from '../lib/merkleTree';
+import { MerkleTree } from '../../lib/merkleTree';
 
 /**
  * Script for creating a merkle tree for testing from a list of UTXOs.
